+++ crates/rust-client/src/transactions/script_builder.rs
@@ -1,4 +1,7 @@
-use alloc::string::String;
+use alloc::{
+    string::{String, ToString},
+    vec::Vec,
+};
 
 use miden_lib::transaction::TransactionKernel;
 use miden_objects::{
@@ -91,7 +94,7 @@ impl AccountInterface {
                         body.push_str(&format!(
                             "
                         push.{asset}
-                        call.wallet::move_asset_to_note dropw 
+                        call.wallet::move_asset_to_note dropw
                         ",
                             asset = prepare_word(&asset.into())
                         ))
@@ -121,11 +124,14 @@ pub(crate) struct TransactionScriptBuilder {
     /// Capabilities of the account for which the script is being built. The capabilities
     /// specify the authentication method and the interfaces exposed by the account.
     account_capabilities: AccountCapabilities,
+    /// The number of blocks in relation to the transaction's reference block after which the
+    /// transaction will expire.
+    expiration_delta: Option<u16>,
 }
 
 impl TransactionScriptBuilder {
-    pub fn new(account_capabilities: AccountCapabilities) -> Self {
-        Self { account_capabilities }
+    pub fn new(account_capabilities: AccountCapabilities, expiration_delta: Option<u16>) -> Self {
+        Self { account_capabilities, expiration_delta }
     }
 
     /// Builds a transaction script which sends the specified notes with the corresponding
@@ -139,10 +145,28 @@ impl TransactionScriptBuilder {
             .interfaces
             .send_note_procedure(self.account_capabilities.account_id, output_notes)?;
 
+        self.build_script_with_sections(vec![send_note_procedure])
+    }
+
+    /// Builds a simple authentication script for the transaction that doesn't send any notes.
+    pub fn build_auth_script(&self) -> Result<TransactionScript, TransactionScriptBuilderError> {
+        self.build_script_with_sections(vec![])
+    }
+
+    /// Builds a transaction script with the specified sections.
+    ///
+    /// The `sections` parameter is a vector of strings, where each string represents a distinct
+    /// part of the script body. The script includes, authentication, and expiration sections are
+    /// automatically added to the script.
+    fn build_script_with_sections(
+        &self,
+        sections: Vec<String>,
+    ) -> Result<TransactionScript, TransactionScriptBuilderError> {
         let script = format!(
-            "{} begin {} {} end",
+            "{} begin {} {} {} end",
             self.script_includes(),
-            send_note_procedure,
+            self.script_expiration(),
+            sections.join(" "),
             self.script_authentication()
         );
 
@@ -152,17 +176,7 @@ impl TransactionScriptBuilder {
         Ok(tx_script)
     }
 
-    /// Builds a simple authentication script for the account that doesn't send any notes.
-    pub fn build_auth_script(&self) -> Result<TransactionScript, TransactionScriptBuilderError> {
-        let script =
-            format!("{} begin {} end", self.script_includes(), self.script_authentication());
-
-        let tx_script = TransactionScript::compile(script, vec![], TransactionKernel::assembler())
-            .map_err(TransactionScriptBuilderError::InvalidTransactionScript)?;
-
-        Ok(tx_script)
-    }
-
+    /// Returns a string with the needed include instructions for the script.
     fn script_includes(&self) -> String {
         let mut includes = String::new();
 
@@ -174,19 +188,27 @@ impl TransactionScriptBuilder {
             },
         }
 
+        if self.expiration_delta.is_some() {
+            includes.push_str("use.miden::tx\n");
+        }
+
         includes
     }
 
+    /// Returns a string with the authentication procedure call for the script.
     fn script_authentication(&self) -> String {
-        let mut body = String::new();
-
         match self.account_capabilities.auth {
-            AuthSecretKey::RpoFalcon512(_) => {
-                body.push_str("call.auth_tx::auth_tx_rpo_falcon512\n");
-            },
+            AuthSecretKey::RpoFalcon512(_) => "call.auth_tx::auth_tx_rpo_falcon512\n".to_string(),
         }
+    }
 
-        body
+    /// Returns a string with the expiration delta update procedure call for the script.
+    fn script_expiration(&self) -> String {
+        if let Some(expiration_delta) = self.expiration_delta {
+            format!("push.{} exec.tx::update_expiration_block_delta\n", expiration_delta)
+        } else {
+            String::new()
+        }
     }
 }